Language detection optimization and demangling improvements ##bin
          swift app  libr_core
old code:   9549      37147
new code:    23       17232

This means that now identifying language in a binary is
from 2 to 400 times faster depending on the binary.

For DVIA iOS app this is 1s less
