+++ libr/bin/Makefile
@@ -21,13 +21,15 @@ include $(LIBR)/magic/deps.mk
 
 STATIC_OBJS=$(addprefix $(LTOP)/bin/p/, $(STATIC_OBJ))
 OBJS=bin.o dbginfo.o bin_ldr.o bin_write.o demangle.o
-OBJS+=dwarf.o filter.o bfile.o obj.o open.o
+OBJS+=dwarf.o filter.o bfile.o obj.o open.o blang.o
 OBJS+=mangling/cxx/cp-demangle.o ${STATIC_OBJS}
 OBJS+=mangling/demangler.o
 OBJS+=mangling/microsoft_demangle.o
 OBJS+=mangling/objc.o mangling/cxx.o mangling/msvc.o
 OBJS+=mangling/swift.o mangling/swift-sd.o
 OBJS+=mangling/dlang.o
+OBJS+=mangling/java.o
+OBJS+=mangling/rust.o
 OBJS+=pdb/pdb_downloader.o pdb/omap.o pdb/stream_pe.o pdb/gdata.o
 OBJS+=pdb/fpo.o pdb/dbi.o pdb/tpi.o pdb/stream_file.o pdb/pdb.o
