+++ libr/bin/mangling/swift-sd.c
@@ -155,7 +155,7 @@ static char *swift_demangle_cmd(const char *s) {
 	return NULL;
 }
 
-R_API char *r_bin_demangle_swift(const char *s, int syscmd) {
+R_API char *r_bin_demangle_swift(const char *s, bool syscmd) {
 #define STRCAT_BOUNDS(x) if (((x) + 2 + strlen (out)) > sizeof (out)) break;
 	char out[1024];
 	int i, len, is_generic = 0;
