+++ libr/bin/meson.build
@@ -4,6 +4,7 @@ files = [
   'dbginfo.c',
   'demangle.c',
   'dwarf.c',
+  'blang.c',
   'filter.c',
   'bfile.c',
   'obj.c',
@@ -98,6 +99,8 @@ files = [
   'mangling/cxx/cp-demangle.c',
   'mangling/demangler.c',
   'mangling/dlang.c',
+  'mangling/rust.c',
+  'mangling/java.c',
   'mangling/microsoft_demangle.c',
   'mangling/msvc.c',
   'mangling/objc.c',
