+++ libr/bin/obj.c
@@ -101,10 +101,39 @@ R_API RBinObject *r_bin_object_new(RBinFile *binfile, RBinPlugin *plugin, ut64 b
 	return o;
 }
 
+static void filter_classes(RBinFile *bf, RList *list) {
+	Sdb *db = sdb_new0 ();
+	RListIter *iter, *iter2;
+	RBinClass *cls;
+	RBinSymbol *sym;
+	r_list_foreach (list, iter, cls) {
+		if (!cls->name) {
+			continue;
+		}
+		int namepad_len = strlen (cls->name) + 32;
+		char *namepad = malloc (namepad_len + 1);
+		if (namepad) {
+			strcpy (namepad, cls->name);
+			r_bin_filter_name (bf, db, cls->index, namepad, namepad_len);
+			free (cls->name);
+			cls->name = namepad;
+			r_list_foreach (cls->methods, iter2, sym) {
+				if (sym->name) {
+					r_bin_filter_sym (bf, db, sym->vaddr, sym);
+				}
+			}
+		} else {
+			eprintf ("Cannot alloc %d byte(s)\n", namepad_len);
+		}
+	}
+	sdb_free (db);
+}
+
 R_API int r_bin_object_set_items(RBinFile *binfile, RBinObject *o) {
 	RBinObject *old_o;
 	RBinPlugin *cp;
 	int i, minlen;
+	bool isSwift = false;
 
 	r_return_val_if_fail (binfile && o && o->plugin, false);
 
@@ -140,6 +169,7 @@ R_API int r_bin_object_set_items(RBinFile *binfile, RBinObject *o) {
 	if (cp->size) {
 		o->size = cp->size (binfile);
 	}
+	// XXX this is expensive because is O(n^n)
 	if (cp->binsym) {
 		for (i = 0; i < R_BIN_SYM_LAST; i++) {
 			o->binsym[i] = cp->binsym (binfile, i);
@@ -174,7 +204,7 @@ R_API int r_bin_object_set_items(RBinFile *binfile, RBinObject *o) {
 				o->symbols->free = r_bin_symbol_free;
 				REBASE_PADDR (o, o->symbols, RBinSymbol);
 				if (bin->filter) {
-					r_bin_filter_symbols (o->symbols);
+					r_bin_filter_symbols (binfile, o->symbols);
 				}
 			}
 		}
@@ -190,7 +220,7 @@ R_API int r_bin_object_set_items(RBinFile *binfile, RBinObject *o) {
 		}
 		REBASE_PADDR (o, o->sections, RBinSection);
 		if (bin->filter) {
-			r_bin_filter_sections (o->sections);
+			r_bin_filter_sections (binfile, o->sections);
 		}
 	}
 	if (bin->filter_rules & (R_BIN_REQ_RELOCS | R_BIN_REQ_IMPORTS)) {
@@ -213,14 +243,15 @@ R_API int r_bin_object_set_items(RBinFile *binfile, RBinObject *o) {
 	if (bin->filter_rules & R_BIN_REQ_CLASSES) {
 		if (cp->classes) {
 			o->classes = cp->classes (binfile);
-			if (r_bin_lang_swift (binfile)) {
+			isSwift = r_bin_lang_swift (binfile);
+			if (isSwift) {
 				o->classes = r_bin_classes_from_symbols (binfile, o);
 			}
 		} else {
 			o->classes = r_bin_classes_from_symbols (binfile, o);
 		}
 		if (bin->filter) {
-			r_bin_filter_classes (o->classes);
+			filter_classes (binfile, o->classes);
 		}
 		// cache addr=class+method
 		if (o->classes) {
@@ -255,7 +286,11 @@ R_API int r_bin_object_set_items(RBinFile *binfile, RBinObject *o) {
 		o->mem = cp->mem (binfile);
 	}
 	if (bin->filter_rules & (R_BIN_REQ_SYMBOLS | R_BIN_REQ_IMPORTS)) {
-		o->lang = r_bin_load_languages (binfile);
+		if (isSwift) {
+			o->lang = R_BIN_NM_SWIFT;
+		} else {
+			o->lang = r_bin_load_languages (binfile);
+		}
 	}
 	binfile->o = old_o;
 	return true;
